Add mongodb database connection
